+++ src/Symfony/Component/HttpFoundation/Tests/Session/Storage/Handler/PdoSessionHandlerTest.php
@@ -98,4 +98,10 @@ class PdoSessionHandlerTest extends \PHPUnit_Framework_TestCase
         $storage->gc(-1);
         $this->assertEquals(0, count($this->pdo->query('SELECT * FROM sessions')->fetchAll()));
     }
+
+    public function testGetConnection()
+    {
+        $storage = new PdoSessionHandler($this->pdo, array('db_table' => 'sessions'), array());
+        $this->assertInstanceOf(get_class($this->pdo), $storage->getConnection());
+    }
 }